GEODE-4751: Address comments.
